* cube:
  bug
  commit

# Conflicts:
#	fbi_cube_base/test/com/finebi/cube/BICubeTestBase.java
#	fbi_cube_generate/src/com/finebi/cube/adapter/BICubeTableAdapter.java
#	fbi_cube_generate/src/com/finebi/cube/gen/oper/BIFieldIndexGenerator.java
#	fbi_cube_generate/src/com/finebi/cube/gen/oper/BIRelationIndexGenerator.java
#	fbi_cube_generate/src/com/finebi/cube/gen/oper/BISourceDataTransport.java

+++ fbi_cube_generate/src/com/finebi/cube/adapter/BICubeTableAdapter.java
@@ -150,8 +150,9 @@ public class BICubeTableAdapter implements ICubeTableService {
     }
 
     @Override
-    public long getTableVersion(BIKey key) {
-        return  primaryTable.getVersion();
+    public int getTableVersion(BIKey key) {
+//        return (int) primaryTable.getCubeLastTime().getTime();
+        return (int) primaryTable.getVersion();
     }
 
     @Override
